requirements_working.txt
accelerate==1.10.1
aiofiles==23.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.13.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.5
bitsandbytes==0.47.0
black==23.12.1
blessed==1.22.0
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
discord.py==2.3.2
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
gpustat==1.1.1
greenlet==3.2.4
griffe==1.13.0
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.26.0
huggingface-hub==0.35.0
idna==3.10
iniconfig==2.1.0
intel-openmp==2021.4.0
isort==5.13.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.10.20
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.20
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-instrumentation==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.6.0
llama-index-llms-openai==0.1.11
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.21
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-postgres==0.1.6
llama-index-workflows==1.3.0
llama-parse==0.4.0
llamaindex-py-client==0.1.19
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.24.4
nvidia-ml-py==13.580.82
openai==1.6.1
orjson==3.11.3
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20221105
pgvector==0.2.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.8.0
pydantic-settings==2.3.0
pydantic_core==2.20.0
Pygments==2.19.2
pyparsing==3.2.4
pypdf==4.3.1
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2023.12.25
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.31
starlette==0.37.2
striprtf==0.0.26
sympy==1.13.1
tbb==2021.11.0
tenacity==8.2.3
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.22.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.16.2
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.17.3
yarl==1.20.1
